requirements.txt
Django>=4.2,<5.1
djangorestframework>=3.14,<3.16
djangorestframework-simplejwt>=5.3,<5.5
django-cors-headers>=4.3,<5
django-filter>=23.5,<26
channels>=4.0,<5
channels-redis>=4.2,<5
django-environ>=0.11,<0.13
django-cleanup>=8.0,<9
drf-spectacular>=0.27,<0.29
django-jazzmin>=3.0,<4
celery>=5.3,<6
redis>=5.0,<6
psycopg2-binary>=2.9,<3
Pillow>=10.0,<12
